Extract `EntityStatus` tests into a different test suite.
